requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
tavily-python==0.3.9
openai==1.54.0
httpx>=0.23.0,<0.28
